Refresh rs-standard.rive from rivescript-js
This re-synchronizes the standard RiveScript brain for the "Try Online"
page from the latest version from rivescript-js, which had benefited
from bug fixes particularly in the Eliza set of replies.

This also fixes the bootstrap.py script to do the least amount of work
as possible in the event that an installation failed for some reason
(e.g. when installing Python dependencies). So now, for example, the
Rophako CMS doesn't need to be deleted and re-cloned every time the
install fails. Instead the script will try to recover from previous
errors, and creates a `.installed` file in the Rophako root as a signal
to itself that the app had been installed correctly.